refactor: bubble 컴포넌트 분리
