+++ src/components/game-room/boardMap/bubble/BubbleContent.tsx
